fix: updated several dependencies + split accent & app locales
