+++ package.json
@@ -51,9 +51,9 @@
     "@emotion/styled": "^11.13.0",
     "@interactivethings/swiss-federal-ci": "^3.0.3",
     "@juggle/resize-observer": "^3.3.1",
-    "@lingui/babel-plugin-lingui-macro": "^5.3.1",
-    "@lingui/core": "^3.13.3",
-    "@lingui/react": "^3.13.3",
+    "@lingui/babel-plugin-lingui-macro": "^5.3.2",
+    "@lingui/core": "^5.3.2",
+    "@lingui/react": "^5.3.2",
     "@mdx-js/react": "^1.6.22",
     "@mui/material": "^5.16.6",
     "@mui/system": "^7.1.0",
@@ -62,11 +62,11 @@
     "@rdfjs/namespace": "^1.1.0",
     "@sentry/nextjs": "^9.22.0",
     "@storybook/nextjs": "^8.6.12",
-    "@tpluscode/rdf-ns-builders": "^1.0.0",
+    "@tpluscode/rdf-ns-builders": "^5.0.0",
     "@tpluscode/rdf-string": "^1.3.3",
     "@tpluscode/sparql-builder": "^2.0.3",
     "@turf/bbox": "^7.2.0",
-    "@turf/centroid": "^6.5.0",
+    "@turf/centroid": "^7.2.0",
     "@turf/turf": "^7.2.0",
     "@types/jsdom": "^21.1.7",
     "@vercel/nft": "^0.27.6",
@@ -86,7 +86,6 @@
     "html2canvas": "^1.4.1",
     "jsdom": "^25.0.1",
     "lodash": "^4.17.21",
-    "make-plural": "^7.0.0",
     "micromark": "^2.11.2",
     "micromark-extension-gfm": "^0.3.3",
     "mitt": "^3.0.1",
@@ -104,7 +103,6 @@
     "sparql-http-client": "^2.3.0",
     "swiss-maps": "^4.6.1",
     "topojson-client": "^3.1.0",
-    "ts-node": "10.7.0",
     "tss-react": "^4.9.16",
     "typescript-lru-cache": "^2.0.0",
     "undici": "^5.25.2",
@@ -115,15 +113,14 @@
   },
   "devDependencies": {
     "@axe-core/playwright": "^4.10.1",
-    "@babel/core": "^7.13.16",
     "@graphql-codegen/cli": "^5.0.2",
     "@graphql-codegen/typescript": "^4.0.9",
     "@graphql-codegen/typescript-operations": "^4.2.3",
     "@graphql-codegen/typescript-resolvers": "^4.2.1",
     "@graphql-codegen/typescript-urql": "^4.0.0",
-    "@lingui/cli": "^3.13.3",
-    "@lingui/macro": "^3.13.3",
-    "@lingui/vite-plugin": "^5.3.0",
+    "@lingui/cli": "^5.3.2",
+    "@lingui/macro": "^5.3.2",
+    "@lingui/vite-plugin": "^5.3.2",
     "@mdx-js/loader": "^1.6.22",
     "@next/bundle-analyzer": "15.0.4",
     "@parcel/watcher": "^2.5.1",
@@ -136,7 +133,6 @@
     "@storybook/test": "^8.6.11",
     "@svgr/cli": "^8.1.0",
     "@types/argparse": "^2.0.10",
-    "@types/clownface": "^1.2.2",
     "@types/d3": "^6.3.0",
     "@types/fs-extra": "^9.0.11",
     "@types/k6": "^0.46.3",
