+++ tsconfig.json
@@ -9,7 +9,7 @@
     "noEmit": true,
     "esModuleInterop": true,
     "module": "esnext",
-    "moduleResolution": "node",
+    "moduleResolution": "bundler",
     "resolveJsonModule": true,
     "isolatedModules": true,
     "jsx": "preserve",
@@ -20,11 +20,7 @@
     },
     "baseUrl": "./"
   },
-  "exclude": [
-    "node_modules",
-    "./src/graphql",
-    "./public/"
-  ],
+  "exclude": ["node_modules", "./src/graphql", "./public/"],
   "include": [
     "*.js",
     "*.ts",
